Refactoring project structure.
