+++ elastic_datashader/helpers/config.py
