+++ elastic_datashader/helpers/drawing.py
