+++ elastic_datashader/helpers/parameters.py
@@ -18,8 +18,8 @@ from elasticsearch import Elasticsearch
 from elasticsearch_dsl import Search, AttrDict, Document
 from elasticsearch.exceptions import NotFoundError, ConflictError
 
-from tms_datashader_api.helpers.timeutil import quantize_time_range, convert_kibana_time
-from tms_datashader_api.helpers.elastic import get_search_base, build_dsl_filter
+from .timeutil import quantize_time_range, convert_kibana_time
+from .elastic import get_search_base, build_dsl_filter
 
 
 def extract_parameters(request):
