+++ elastic_datashader/helpers/timeutil.py
